The plugin now generates lists of vulnerabilities
It turns out that leiningen already indirectly included lucene through maven-index, so the solution was to launch the analysis of project classpath in a separate process with only plugin as dependency.

Lots of code moving around in this change as leiningen support would not be available in isolated subprocess. 99% of this is from Doug's initial commit. The main difference grabbing essential info project-classpath project-name output-format output-directory before forking new subprocess.